requirements.txt
async-timeout==4.0.3
grpcio==1.63.0
grpcio-tools==1.63.0
pika==1.3.2
protobuf==5.26.1
redis==5.0.4
